Rebuilt controllers using the new input system
Everything related to the player controller was rebuilt to allow scalability for the game. Every player action has his own class (component, script) to be used from the head class called "MyPlayerInputManager", Added the word 'My' to distinguish from the built in "PlayerInputManager" component. MyPlayerInputManager receives the input from the "PlayerInput" Component and deal with it accordingly.